algorithm_params.edge_construction_params.intersected_edge_coeff is now double, default value in config.info.template is 1.0. We construct an intersected edge if num_individual_sms < coeff*num_shared_shms

+++ src/antevolo/antevolo_config.hpp
@@ -32,7 +32,7 @@ namespace antevolo {
             };
 
             struct EdgeConstructionParams {
-                size_t intersected_edge_coeff;
+                double intersected_edge_coeff;
                 size_t min_num_intersected_v_shms;
             };
